requirements.txt
altair==5.5.0
APScheduler==3.11.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.48.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.41
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
